取消Preview预览参数，只保留groupKey合并src预览
