+++ examples/group.tsx
@@ -1,0 +1,42 @@
+import * as React from 'react';
+import Image from '../src';
+import '../assets/index.less';
+
+export default function Urls() {
+  return (
+    <div>
+      <Image
+        groupKey="preview"
+        src="https://zos.alipayobjects.com/rmsportal/jkjgkEfvpUPVyRjUImniVslZfWPnJuuZ.png"
+        width={200}
+        style={{
+          marginRight: 24,
+        }}
+      />
+      <Image
+        groupKey="preview"
+        src="https://gw.alipayobjects.com/mdn/rms_08e378/afts/img/A*P0S-QIRUbsUAAAAAAAAAAABkARQnAQ"
+        width={200}
+        style={{
+          marginRight: 24,
+        }}
+      />
+      <Image
+        groupKey="preview"
+        src="https://gw.alipayobjects.com/mdn/rms_08e378/afts/img/A*ngiJQaLQELEAAAAAAAAAAABkARQnAQ"
+        width={200}
+        style={{
+          marginRight: 24,
+        }}
+      />
+      <Image
+        groupKey="preview"
+        src="https://gw.alipayobjects.com/mdn/rms_08e378/afts/img/A*NZuwQp_vcIQAAAAAAAAAAABkARQnAQ"
+        width={200}
+        style={{
+          marginRight: 24,
+        }}
+      />
+    </div>
+  );
+}
